+++ blockfrost-api/src/Blockfrost/API/Cardano.hs
@@ -12,6 +12,7 @@ module Blockfrost.API.Cardano
   , module Blockfrost.API.Cardano.Metadata
   , module Blockfrost.API.Cardano.Network
   , module Blockfrost.API.Cardano.Pools
+  , module Blockfrost.API.Cardano.Scripts
   , module Blockfrost.API.Cardano.Transactions
   ) where
 
@@ -24,4 +25,5 @@ import Blockfrost.API.Cardano.Ledger
 import Blockfrost.API.Cardano.Metadata
 import Blockfrost.API.Cardano.Network
 import Blockfrost.API.Cardano.Pools
+import Blockfrost.API.Cardano.Scripts
 import Blockfrost.API.Cardano.Transactions