+++ blockfrost-api/src/Blockfrost/Types/Cardano.hs
@@ -10,6 +10,7 @@ module Blockfrost.Types.Cardano
   , module Blockfrost.Types.Cardano.Metadata
   , module Blockfrost.Types.Cardano.Network
   , module Blockfrost.Types.Cardano.Pools
+  , module Blockfrost.Types.Cardano.Scripts
   , module Blockfrost.Types.Cardano.Transactions
   ) where
 
@@ -22,4 +23,5 @@ import Blockfrost.Types.Cardano.Genesis
 import Blockfrost.Types.Cardano.Metadata
 import Blockfrost.Types.Cardano.Network
 import Blockfrost.Types.Cardano.Pools
+import Blockfrost.Types.Cardano.Scripts
 import Blockfrost.Types.Cardano.Transactions